+++ src/client.ts
@@ -1,55 +1,148 @@
-import http from 'http';
-
-import { IntegrationProviderAuthenticationError } from '@jupiterone/integration-sdk-core';
-
+import fetch, { Response } from 'node-fetch';
+import { retry } from '@lifeomic/attempt';
+import {
+  IntegrationProviderAPIError,
+  IntegrationProviderAuthenticationError,
+} from '@jupiterone/integration-sdk-core';
 import { IntegrationConfig } from './config';
-import { AcmeUser, AcmeGroup } from './types';
+import {
+  DataStaxDatabase,
+  DataStaxAccessList,
+  DataStaxUsers,
+  DataStaxRole,
+  DataStaxUser,
+} from './types';
 
 export type ResourceIteratee<T> = (each: T) => Promise<void> | void;
 
-/**
- * An APIClient maintains authentication state and provides an interface to
- * third party data APIs.
- *
- * It is recommended that integrations wrap provider data APIs to provide a
- * place to handle error responses and implement common patterns for iterating
- * resources.
- */
 export class APIClient {
   constructor(readonly config: IntegrationConfig) {}
 
-  public async verifyAuthentication(): Promise<void> {
-    // TODO make the most light-weight request possible to validate
-    // authentication works with the provided credentials, throw an err if
-    // authentication fails
-    const request = new Promise<void>((resolve, reject) => {
-      http.get(
-        {
-          hostname: 'localhost',
-          port: 443,
-          path: '/api/v1/some/endpoint?limit=1',
-          agent: false,
-          timeout: 10,
+  private perPage = 100;
+  private baseUri = `https://api.astra.datastax.com`;
+  private withBaseUri = (path: string) => `${this.baseUri}${path}`;
+
+  private checkStatus = (response: Response) => {
+    if (response.ok) {
+      return response;
+    } else {
+      throw new IntegrationProviderAPIError(response);
+    }
+  };
+
+  private async request(
+    uri: string,
+    method: 'GET' | 'HEAD' = 'GET',
+  ): Promise<any> {
+    try {
+      // Handle rate-limiting
+      const response = await retry(
+        async () => {
+          const res: Response = await fetch(uri, {
+            method,
+            headers: {
+              Authorization: `Bearer ${this.config.token}`,
+            },
+          });
+          return this.checkStatus(res);
         },
-        (res) => {
-          if (res.statusCode !== 200) {
-            reject(new Error('Provider authentication failed'));
-          } else {
-            resolve();
-          }
+        {
+          delay: 5000,
+          maxAttempts: 10,
+          handleError: (err, context) => {
+            if (
+              err.statusCode !== 429 ||
+              ([500, 502, 503].includes(err.statusCode) &&
+                context.attemptNum > 1)
+            )
+              context.abort();
+          },
         },
       );
-    });
-
-    try {
-      await request;
+      return response.json();
     } catch (err) {
-      throw new IntegrationProviderAuthenticationError({
-        cause: err,
-        endpoint: 'https://localhost/api/v1/some/endpoint?limit=1',
+      throw new IntegrationProviderAPIError({
+        endpoint: uri,
         status: err.status,
         statusText: err.statusText,
       });
+    }
+  }
+
+  private async paginatedRequest<T>(
+    uri: string,
+    method: 'GET' | 'HEAD' = 'GET',
+    iteratee: ResourceIteratee<T>,
+    iterateeKey: string,
+  ): Promise<void> {
+    try {
+      let nextUri = null;
+      do {
+        const response = await this.request(nextUri || uri, method);
+        nextUri =
+          response.length > 0
+            ? `${uri}&starting_after=${
+                response[response.length - 1][iterateeKey]
+              }`
+            : response['@pagination']?.next;
+        for (const item of response) {
+          await iteratee(item);
+        }
+      } while (nextUri);
+    } catch (err) {
+      throw new IntegrationProviderAPIError({
+        cause: new Error(err.message),
+        endpoint: uri,
+        status: err.statusCode,
+        statusText: err.message,
+      });
+    }
+  }
+
+  public async verifyAuthentication(): Promise<void> {
+    const uri = this.withBaseUri('/v2/access-lists');
+    try {
+      await this.request(uri);
+    } catch (err) {
+      throw new IntegrationProviderAuthenticationError({
+        cause: err,
+        endpoint: uri,
+        status: err.status,
+        statusText: err.statusText,
+      });
+    }
+  }
+
+  /**
+   * Iterates each database resource in the provider.
+   *
+   * @param iteratee receives each resource to produce entities/relationships
+   */
+  public async iterateDatabases(
+    iteratee: ResourceIteratee<DataStaxDatabase>,
+  ): Promise<void> {
+    await this.paginatedRequest<DataStaxDatabase>(
+      this.withBaseUri(`/v2/databases?limit=${this.perPage}`),
+      'GET',
+      iteratee,
+      'id',
+    );
+  }
+
+  /**
+   * Iterates each access list resource in the provider.
+   *
+   * @param iteratee receives each resource to produce entities/relationships
+   */
+  public async iterateAccessLists(
+    iteratee: ResourceIteratee<DataStaxAccessList>,
+  ): Promise<void> {
+    const accessLists: DataStaxAccessList[] = await this.request(
+      this.withBaseUri(`/v2/access-lists`),
+      'GET',
+    );
+    for (const accessList of accessLists) {
+      await iteratee(accessList);
     }
   }
 
@@ -59,62 +152,31 @@ export class APIClient {
    * @param iteratee receives each resource to produce entities/relationships
    */
   public async iterateUsers(
-    iteratee: ResourceIteratee<AcmeUser>,
+    iteratee: ResourceIteratee<DataStaxUser>,
   ): Promise<void> {
-    // TODO paginate an endpoint, invoke the iteratee with each record in the
-    // page
-    //
-    // The provider API will hopefully support pagination. Functions like this
-    // should maintain pagination state, and for each page, for each record in
-    // the page, invoke the `ResourceIteratee`. This will encourage a pattern
-    // where each resource is processed and dropped from memory.
-
-    const users: AcmeUser[] = [
-      {
-        id: 'acme-user-1',
-        name: 'User One',
-      },
-      {
-        id: 'acme-user-2',
-        name: 'User Two',
-      },
-    ];
-
-    for (const user of users) {
+    const users: DataStaxUsers = await this.request(
+      this.withBaseUri(`/v2/organizations/users`),
+      'GET',
+    );
+    for (const user of users.Users) {
       await iteratee(user);
     }
   }
 
   /**
-   * Iterates each group resource in the provider.
+   * Iterates each role resource in the provider.
    *
    * @param iteratee receives each resource to produce entities/relationships
    */
-  public async iterateGroups(
-    iteratee: ResourceIteratee<AcmeGroup>,
+  public async iterateRoles(
+    iteratee: ResourceIteratee<DataStaxRole>,
   ): Promise<void> {
-    // TODO paginate an endpoint, invoke the iteratee with each record in the
-    // page
-    //
-    // The provider API will hopefully support pagination. Functions like this
-    // should maintain pagination state, and for each page, for each record in
-    // the page, invoke the `ResourceIteratee`. This will encourage a pattern
-    // where each resource is processed and dropped from memory.
-
-    const groups: AcmeGroup[] = [
-      {
-        id: 'acme-group-1',
-        name: 'Group One',
-        users: [
-          {
-            id: 'acme-user-1',
-          },
-        ],
-      },
-    ];
-
-    for (const group of groups) {
-      await iteratee(group);
+    const roles: DataStaxRole[] = await this.request(
+      this.withBaseUri(`/v2/organizations/roles`),
+      'GET',
+    );
+    for (const role of roles) {
+      await iteratee(role);
     }
   }
 }