+++ src/steps/database/index.test.ts
@@ -1,0 +1,21 @@
+import { executeStepWithDependencies } from '@jupiterone/integration-sdk-testing';
+import { buildStepTestConfigForStep } from '../../../test/config';
+import { Recording, setupProjectRecording } from '../../../test/recording';
+import { Steps } from '../constants';
+
+// See test/README.md for details
+let recording: Recording;
+afterEach(async () => {
+  await recording.stop();
+});
+
+test('fetch-databases', async () => {
+  recording = setupProjectRecording({
+    directory: __dirname,
+    name: 'fetch-databases',
+  });
+
+  const stepConfig = buildStepTestConfigForStep(Steps.DATABASES);
+  const stepResult = await executeStepWithDependencies(stepConfig);
+  expect(stepResult).toMatchStepMetadata(stepConfig);
+});